tools/v2/python/requirements.txt
numpy<2
pygame
pyopengl
pyopengl_accelerate
pyserial
roboticstoolbox-python